tests/requirements.txt
# Test Dependencies for BEE Engagement Events Testing Suite
# Module: Core Engagement
# Milestone: 1 · Data Backbone

# Database testing
psycopg2-binary>=2.9.0
pytest>=7.0.0

# API testing
requests>=2.28.0

# Environment management
python-dotenv>=0.19.0

# Performance testing and concurrency
asyncio>=3.4.3

# Data manipulation and analysis
statistics  # Built-in module

# Optional: Enhanced testing features
pytest-html>=3.1.0  # For HTML test reports
pytest-cov>=4.0.0   # For coverage reports
pytest-xdist>=2.5.0 # For parallel test execution 